requirements.txt
Django==2.1.1
django-cors-headers==2.4.0
django-environ==0.4.5
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
psycopg2==2.7.5
PyJWT==1.6.4
pytz==2018.5
faker===0.9.2
coverage==3.6